feat: consolidate Solidity interfaces into dedicated interfaces.rs files
- Create consolidated interfaces.rs files for ERC20, ERC721, ERC1155, and proxy contracts
- Separate callable interfaces (stylus_proc::sol_interface) from non-callable (alloy_sol_types::sol)
- Move scattered interface definitions to appropriate centralized locations:
  * ERC20: Erc20Interface, IErc20MetadataInterface, IERC3156FlashBorrower, IERC20, IERC1363
  * ERC721: Erc721Interface, IErc721ReceiverInterface
  * ERC1155: IErc1155ReceiverInterface
  * Proxy: IBeaconInterface, Erc1822ProxiableInterface
- Update imports throughout codebase to use consolidated interfaces
- Maintain backward compatibility and preserve all existing functionality

Resolves #772
