+++ contracts/src/token/erc1155/mod.rs
@@ -21,13 +21,15 @@ use crate::utils::{
 };
 
 pub mod extensions;
+pub mod interfaces;
 pub mod receiver;
 pub mod utils;
 
 pub use receiver::{
-    IErc1155Receiver, IErc1155ReceiverInterface, BATCH_TRANSFER_FN_SELECTOR,
+    IErc1155Receiver, BATCH_TRANSFER_FN_SELECTOR,
     SINGLE_TRANSFER_FN_SELECTOR,
 };
+pub use interfaces::IErc1155ReceiverInterface;
 pub use sol::*;
 #[cfg_attr(coverage_nightly, coverage(off))]
 mod sol {
