+++ contracts/src/token/erc1155/receiver.rs
@@ -2,11 +2,11 @@
 //! ERC-1155 token transfers.
 #![allow(missing_docs)]
 #![cfg_attr(coverage_nightly, coverage(off))]
-use alloc::{vec, vec::Vec};
+use alloc::vec::Vec;
 
 use alloy_primitives::{aliases::B32, Address, U256};
 use openzeppelin_stylus_proc::interface_id;
-use stylus_sdk::{abi::Bytes, function_selector, prelude::*};
+use stylus_sdk::{abi::Bytes, function_selector};
 
 use crate::utils::introspection::erc165::IErc165;
 
@@ -30,33 +30,6 @@ pub const BATCH_TRANSFER_FN_SELECTOR: B32 = B32::new(function_selector!(
     Vec<U256>,
     Bytes
 ));
-
-sol_interface! {
-    /// [`super::Erc1155`] token receiver Solidity interface.
-    ///
-    /// Check [`super::IErc1155Receiver`] trait for more details.
-    interface IErc1155ReceiverInterface {
-        /// See [`super::IErc1155Receiver::on_erc1155_received`].
-        #[allow(missing_docs)]
-        function onERC1155Received(
-            address operator,
-            address from,
-            uint256 id,
-            uint256 value,
-            bytes calldata data
-        ) external returns (bytes4);
-
-        /// See [`super::IErc1155Receiver::on_erc1155_batch_received`].
-        #[allow(missing_docs)]
-        function onERC1155BatchReceived(
-            address operator,
-            address from,
-            uint256[] calldata ids,
-            uint256[] calldata values,
-            bytes calldata data
-        ) external returns (bytes4);
-    }
-}
 
 /// Interface that must be implemented by smart contracts in order to receive
 /// ERC-1155 token transfers.
