+++ contracts/src/token/erc20/mod.rs
@@ -22,6 +22,7 @@ use crate::utils::{
 
 pub mod extensions;
 pub mod interface;
+pub mod interfaces;
 pub mod utils;
 
 pub use sol::*;
